[MIG] website_field_autocomplete: Migration to 17.0
