+++ website_field_autocomplete/__manifest__.py
@@ -4,7 +4,7 @@
 {
     "name": "Website Field - AutoComplete",
     "summary": "Provides an autocomplete field for Website on any model",
-    "version": "14.0.1.0.0",
+    "version": "17.0.1.0.0",
     "category": "Website",
     "website": "https://github.com/OCA/website",
     "author": "LasLabs, Odoo Community Association (OCA)",
@@ -14,9 +14,13 @@
     "depends": [
         "website",
     ],
-    "data": [
-        "views/assets.xml",
-    ],
+    "assets": {
+        "web.assets_frontend": [
+            "website_field_autocomplete/static/src/lib/jquery.ui/jquery-ui.css",
+            "website_field_autocomplete/static/src/lib/jquery.ui/jquery-ui.js",
+            "website_field_autocomplete/static/src/js/field_autocomplete.esm.js",
+        ],
+    },
     "demo": [
         "demo/field_autocomplete_demo.xml",
     ],
