+++ website_field_autocomplete/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_website_field_autocomplete
