--- website_field_autocomplete/views/assets.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<!--
-    Copyright 2016 LasLabs Inc.
-    @license AGPL-3 or later (http://www.gnu.org/licenses/agpl.html).
--->
-<odoo>
-    <template
-        id="assets_autocomplete"
-        name="website_field_autocomplete Assets"
-        inherit_id="website.assets_frontend"
-    >
-        <xpath expr="." position="inside">
-            <script
-                src="/website_field_autocomplete/static/src/js/field_autocomplete.js"
-            />
-        </xpath>
-    </template>
-</odoo>
